WebApp/requirements.txt
flask==1.1.2
redis==3.5.3
pybase62==0.4.3